add multiple images

+++ src/components/dashboard/ItemForm.tsx
@@ -12,35 +12,42 @@ interface Props {
 
 export const ItemForm = ({ item, onSubmit, isCreation }: Props) => {
   const [message, setMessage] = useState<string | null>(null);
-  const [file, setFile] = useState<File>();
+  const [files, setFiles] = useState<FileList | null>();
 
   async function handleSubmit(e: any) {
     e.preventDefault();
 
     const formData = new FormData(e.target);
 
-    if (file)
+    if (files)
     {
       try {
+
+        if (!isCreation && item)
+        {
+          const dataToDelete = new FormData()
+          for (const fileName of item.images) {
+            dataToDelete.append('fileNames', fileName);
+          };
+          await fetch('/api/upload', {
+            method: 'DELETE',
+            body: dataToDelete
+          });
+        }
+        
         const data = new FormData();
-        data.set('file', file);
-  
+
+        for (const file of Array.from(files)) {
+          data.append('files', file);
+        }
+
         const post = await fetch('/api/upload', {
           method: 'POST',
           body: data
         });
   
-        const response: { fileName: string } = await post.json();
-  
-        onSubmit(convertFormDataToWriteItem(formData, response.fileName));
-
-        if (!isCreation)
-        {
-          await fetch('/api/upload', {
-            method: 'DELETE',
-            body: JSON.stringify({ fileName: item?.images[0] }),
-          });
-        }
+        const response: { fileNames: string[] } = await post.json();
+        onSubmit(convertFormDataToWriteItem(formData, response.fileNames));
 
         // Handle the error
         if (!post.ok) throw new Error(await post.text())
@@ -49,19 +56,18 @@ export const ItemForm = ({ item, onSubmit, isCreation }: Props) => {
         console.error(e);
       }
     } else {
-      onSubmit(convertFormDataToWriteItem(formData, item?.images[0] ?? ""));
+      onSubmit(convertFormDataToWriteItem(formData, item?.images ?? [""]));
     }
     setMessage(isCreation ? "L'article a bien été créé." : "L'article a bien été modifié.")
   }
 
-  function convertFormDataToWriteItem(formData: FormData, filename : string): WriteItem {
+  function convertFormDataToWriteItem(formData: FormData, filenames : string[]): WriteItem {
     const title = formData.get('title') as string;
     const description = formData.get('description') as string | null;
     const price = new Prisma.Decimal(parseFloat(formData.get('price') as string));
     const type = formData.get('type') as Type;
     const published = formData.get('published') === 'on';
-    const images = [ filename ];
-    // const images = (formData.getAll('images') as string[]).filter(Boolean);
+    const images = filenames;
 
     const writeItem: WriteItem = {
       title,
@@ -107,14 +113,15 @@ export const ItemForm = ({ item, onSubmit, isCreation }: Props) => {
             <input type="checkbox" name="published" defaultChecked={item?.published} />
           </label>
         </div>
-        <div className="flex flex-col md:flex-row space-x-3">
+        <div className="flex flex-col space-x-3">
           <p>Importer une image :</p>
-          <div className="p-3 min-h-50 w-full sm:w-1/2 md:w-1/2 lg:w-1/3 xl:w-1/4 2xl:w-1/4">
-            <img className="object-cover h-24 w-24 sm:h-48 sm:w-48" src={"/uploads/" + item?.images[0]} alt="" />
+          <div className="flex flew-wrap p-3 min-h-50 w-full sm:w-1/2 md:w-1/2 lg:w-1/3 xl:w-1/4 2xl:w-1/4">
+            {item?.images.map((image) => (
+              <img className="m-2 object-cover h-24 w-20 sm:h-52 sm:w-40" src={"/uploads/" + image} alt="" />
+            ))}
           </div>
           <label className="flex flex-row space-x-3">
-            {/* <input name="file" ref={inputFileRef} type="file" required={isCreation}/> */}
-            <input type="file" name="file" accept="image/*" onChange={(e) => setFile(e.target.files?.[0])} required={isCreation} />
+            <input type="file" name="file" accept="image/*" multiple onChange={(e) => setFiles(e.target.files)} required={isCreation} />
           </label>
         </div>
         <button className="bg-orange-200 border border-orange-600 p-4" type="submit">Enregistrer l&lsquo;article</button>
